requirements.txt
fastapi
uvicorn[standard]
numpy==1.26.4
pillow
insightface==0.7.3
onnxruntime